Refactor decimal helper functions into separate file.
